fix: move step flows to utilities
