+++ packages/ibm-products-web-components/src/utilities/step-flow/_story-assets/step-full-page.scss
