+++ packages/ibm-products-web-components/src/utilities/step-flow/_story-assets/step-full-page.ts
@@ -19,7 +19,7 @@ import '../step-group';
 import {
   registerFocusableContainers,
   trapFocus,
-} from '../../../utilities/manageFocusTrap/manageFocusTrap';
+} from '../../manageFocusTrap/manageFocusTrap';
 
 interface FormStateType {
   email?: string;
