+++ packages/ibm-products-web-components/src/utilities/step-flow/_story-assets/step-side-panel.scss
