+++ packages/ibm-products-web-components/src/utilities/step-flow/_story-assets/step-tearsheet.ts
@@ -18,7 +18,7 @@ import '../../../components/tearsheet/index.js';
 import {
   registerFocusableContainers,
   trapFocus,
-} from '../../../utilities/manageFocusTrap/manageFocusTrap';
+} from '../../manageFocusTrap/manageFocusTrap';
 
 interface FormStateType {
   email?: string;
