+++ packages/ibm-products-web-components/src/utilities/step-flow/step-flow.stories.ts
@@ -64,5 +64,5 @@ export const SidePanel = {
 };
 
 export default {
-  title: 'Patterns/Step flows',
+  title: 'Utilities/Step flows',
 };
