+++ packages/ibm-products-web-components/src/utilities/step-flow/story-styles.scss
