+++ packages/ibm-products/src/global/js/utils/StepFlow/_storybook-styles.scss
