+++ packages/ibm-products/src/global/js/utils/StepFlow/story-assets/SidePanel/Example.tsx
