+++ packages/ibm-products/src/global/js/utils/StepFlow/story-assets/SidePanel/_step-side-panel.scss
