+++ packages/ibm-products/src/global/js/utils/StepFlow/story-assets/Tearsheet/Example.tsx
